+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+
+## [0.4.0] - 2018-07-13
 ### Added
 - ServiceProviderConfiguration class handles configuration for a specific host
 - SsoResponse class
@@ -42,7 +44,8 @@
 - Coveralls Integration
 - Rubygems version badge in README
 
-[Unreleased]: https://github.com/italia/spid-ruby/compare/v0.3.1...HEAD
+[Unreleased]: https://github.com/italia/spid-ruby/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/italia/spid-ruby/compare/v0.3.1...v0.4.0
 [0.3.1]: https://github.com/italia/spid-ruby/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/italia/spid-ruby/compare/v0.2.2...v0.3.0
 [0.2.2]: https://github.com/italia/spid-ruby/compare/v0.2.1...v0.2.2